refactoring: moved bundler and context creation to sepearate files
