+++ src/index.js
@@ -3,6 +3,7 @@ import { validateFromFile } from './validate';
 import cli from './cli';
 
 export { validate, validateFromFile } from './validate';
+export { bundle } from './bundle';
 export default validateFromFile;
 
 if (require.main === module) {
